Add the printouts for the uplink case
We aren't going all the way to correct errors, for a couple of reasons.

Reason 1: It turned out too difficult to implement with all the
syndrome calculations, Berlekamp-Massey algorithm, and whatnot.
You should be using dump978 if you want all that.

Reason 2: It turned out that there aren't any correctable errors
in ruat. Our demodulator is so crude, that any bits which manage to
squeeze through are always error-free.